Change source code struct
